+++ CHANGELOG.md
@@ -404,6 +404,7 @@ Bug fixes, improvements on PDF export function, added major support for the webp
 - [x] Added full pdf support for markdown rendering result for image caption and description. 
 - [x] Added metadata injection to the pdf, which retains the metadata for the figures. It will also add detailed citation data to the 
 - [x] Added better spacing for the image export of PDF, preventing images to become crowd in many cases 
+- [x] Added batch markdown export command `Sync repository to the website notes folder`.
 
 🐛 Fixed Bugs : 
 - [x] fixed [#162](https://github.com/FRIEDparrot/obsidian-equation-citator/issues/162), viewState refresh correctly when switch files